move golang e2e our of source tree and into test/e2e tree
